# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: OctoFHIR API
  description: |
    OctoFHIR is a high-performance FHIR R4/R5/R6 server built in Rust.

    This API reference covers:
    - FHIR REST API (CRUD, search, operations)
    - OAuth 2.0 / SMART on FHIR authentication
    - Admin API for user and configuration management
    - GraphQL endpoint
  version: 1.0.0
  contact:
    name: OctoFHIR Support
    url: https://github.com/octofhir/server-rs
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
servers:
  - url: http://localhost:8888
    description: Local development server
  - url: https://fhir.example.com
    description: Production server (example)
...
paths:
...
  /auth/logout:
    post:
      tags: [OAuth]
      summary: Logout
      description: Revoke tokens and clear session
      operationId: logout
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Logout successful
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: OAuth 2.0 Bearer token
...
  schemas:
...
    OperationOutcome:
      type: object
      required: [resourceType, issue]
      properties:
        resourceType:
          type: string
          enum: [OperationOutcome]
        issue:
          type: array
          items:
            type: object
            required: [severity, code]
            properties:
              severity:
                type: string
                enum: [fatal, error, warning, information]
              code:
                type: string
              diagnostics:
                type: string
              location:
                type: array
                items:
                  type: string
...
  responses:
...
    Unauthorized:
      description: Unauthorized - missing or invalid token
      content:
        application/fhir+json:
          schema:
            $ref: '#/components/schemas/OperationOutcome'
...
security:
  - bearerAuth: []